+++ src/main/drivers/accgyro/accgyro.h
@@ -46,8 +46,8 @@ typedef struct gyroDev_s {
     sensorGyroUpdateFuncPtr updateFn;
     extiCallbackRec_t exti;
     float scale;                                        // scalefactor
-    int16_t gyroADCRaw[XYZ_AXIS_COUNT];
-    int16_t gyroZero[XYZ_AXIS_COUNT];
+    float gyroADCRaw[XYZ_AXIS_COUNT];
+    float gyroZero[XYZ_AXIS_COUNT];
     uint8_t imuSensorToUse;
     uint8_t lpf;                                        // Configuration value: Hardware LPF setting
     uint32_t requestedSampleIntervalUs;                 // Requested sample interval
